% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{ajuste_modelos_distance_hn}
\alias{ajuste_modelos_distance_hn}
\title{Ajusta a funcao de deteccao Half-Normal com diferentes termos de ajuste}
\usage{
ajuste_modelos_distance_hn(
  dados,
  lista_termos_ajuste = list(`Sem termo` = NULL, Cosseno = "cos", `Hermite polinomial` =
    "herm"),
  truncamento = NULL,
  formula = ~1
)
}
\arguments{
\item{truncamento}{}
}
\description{
Ajusta a funcao de deteccao Half-Normal com diferentes termos de ajuste
}
