% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{ajuste_modelos_distance_hr}
\alias{ajuste_modelos_distance_hr}
\title{Title}
\usage{
ajuste_modelos_distance_hr(
  dados,
  lista_termos_ajuste = list(`Sem termo` = NULL, Cosseno = "cos", `Polinomial simples` =
    "poly"),
  truncamento = NULL,
  formula = ~1
)
}
\arguments{
\item{truncamento}{}
}
\description{
Title
}
