% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{ajuste_modelos_distance_unif}
\alias{ajuste_modelos_distance_unif}
\title{Title}
\usage{
ajuste_modelos_distance_unif(
  dados,
  lista_termos_ajuste = list(Cosseno = "cos", `Polinomial simples` = "poly"),
  truncamento = NULL,
  formula = ~1
)
}
\arguments{
\item{truncamento}{}
}
\description{
Title
}
