% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{carregar_dados_brutos_rds}
\alias{carregar_dados_brutos_rds}
\title{Carrega os dados brutos originais em formato .rds}
\usage{
carregar_dados_brutos_rds(
  dados = readr::read_rds(file = paste0(here::here(), "/data-raw/dados_brutos.rds"))
)
}
\arguments{
\item{dados}{recebe o caminho para carregar um arquivo .rds. Por configuracao, carrega os dados brutos dos Projeto Monitora Componente Florestal a partir do formato .rds}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo os dados brutos do Projeto Monitora Componente Florestal.
}
\description{
A funcao \code{carregar_dados_brutos_rds()} carrega os dados brutos em formato .rds.
}
\details{
A funcao \code{carregar_dados_brutos_rds()} carrega e disponibiliza ao usario os dados brutos do Projeto Monitora Componente Florestal, contendo  as infromacoes sobre as amostragens por distancia (\emph{distance samplig}) de aves e pequenos e medios mamiferos realizadas desde 2014 em 40 Unidades de Conservacao do Brasil.
}
\examples{
\dontrun{carregar_dados_brutos_rds()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
