% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carregar_dados_brutos_xlsx.R
\name{carregar_dados_brutos_xlsx}
\alias{carregar_dados_brutos_xlsx}
\title{Carrega os dados brutos originais em formato .xlsx}
\usage{
carregar_dados_brutos_xlsx(
  dados = "data-raw/monitora_masto_aves_2023_04_04.xlsx",
  planilha = "dados brutos"
)
}
\arguments{
\item{dados}{recebe o caminho para carregar um arquivo .xlsx. Por configuracao, carrega os dados brutos dos Projeto Monitora Componente Florestal \code{"monitora_masto_aves_2023_04_04.xlsx"}. A versao dos dados brutos corresponde ao arquivo "Planilha Oficial consolidada de Masto-aves 2014-21 Validada CEMAVE CPB CENAP.xlsx"}

\item{planilha}{nome ou número da planilha no arquivo .xlsx.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo os dados brutos do Projeto Monitora Componente Florestal a partir do formato .xlsx.
}
\description{
A funcao \code{carregar_dados_brutos_xlsx()} carrega o arquivo da planilha de dados originais, \code{"monitora_masto_aves_2023_04_04.xlsx"} e gera um novo arquivo \code{dados_brutos.rds}.
}
\details{
A funcao \code{carregar_dados_brutos_xlsx()} carrega e disponibiliza ao usario os dados brutos do Projeto Monitora Componente Florestal, contendo  as infromacoes sobre as amostragens por distancia (\emph{distance samplig}) de aves e pequenos e medios mamiferos realizadas desde 2014 em 40 Unidades de Conservacao do Brasil.
}
\examples{
\dontrun{carregar_dados_brutos_xlsx()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
