% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{carregar_dados_completos}
\alias{carregar_dados_completos}
\title{Carrega os dados brutos selecionando, trasformando e renomeando as colunas}
\usage{
carregar_dados_completos(
  dados = readr::read_rds(file = paste0(here::here(), "/inst/extdata/dados_brutos.rds"))
)
}
\arguments{
\item{dados}{recebe o caminho para carregar um arquivo .rds. Por configuracao, carrega os dados brutos dos Projeto Monitora Componente Florestal a partir do formato .rds}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo uma selecao de colunas transformadas e renomeadas a partir dos dados brutos do Projeto Monitora Componente Florestal.
}
\description{
A funcao \code{carregar_dados_completos()} carrega os dados brutos em formato .rds, selecionando, transformando e renomeando suas colunas.
}
\details{
A funcao \code{carregar_dados_completos()} disponibiliza ao usario os dados brutos do Projeto Monitora Componente Florestal, selecionando, transformando e renomeando suas colunas. (Fornecer mais detalhes sobre as transformacoes operadas).
}
\examples{
\dontrun{carregar_dados_completos()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
