% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{carregar_dados_filtrados}
\alias{carregar_dados_filtrados}
\title{Carrega os dados filtrados a partir da UC's e da especie}
\usage{
carregar_dados_filtrados(
  dados = readr::read_rds(file = paste0(here::here(), "/data/dados_completos.rds")),
  nome_uc = "Resex Tapajos-Arapiuns",
  nome_sp = "Dasyprocta croconota"
)
}
\arguments{
\item{dados}{recebe o caminho para carregar um arquivo .rds. Por configuracao, carrega os dados filtrados pela Resex Tapajos-Arapiuns e especie \emph{Dsyprocta croconota} do Projeto Monitora Componente Florestal a partir doa arquivo \code{"dados_completos.rds"}.}

\item{nome_uc}{description}

\item{nome_sp}{descricao}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo os dados filtrados do Projeto Monitora Componente Florestal.
}
\description{
A funcao \code{carregar_dados_filtrados()} carrega o arquivo de dados completos, \code{"dados_completos.rds"} e gera um novo arquivo \code{dados_filtrados.rds}.
}
\details{
A funcao \code{carregar_dados_filtrados()} carrega e disponibiliza ao usario os dados filtrados do Projeto Monitora Componente Florestal, contendo ...
}
\examples{
\dontrun{carregar_dados_filtrados()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
