% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_ano_uc.R
\name{contar_n_ano_uc}
\alias{contar_n_ano_uc}
\title{Número total de anos em que cada Unidade de Consevarção foi amostrada}
\usage{
contar_n_ano_uc(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pelas funcoes \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_floresta}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo três colunas, \code{nome_uc} que contem os nomes completos das Unidades de Conservação, \code{nome_uc_abv} que contem os nomes abreviados das Unidades de Conservação e \code{n_anos} que contem o número de anos em que cada Unidade de Conservação foi amostrada.
}
\description{
A função \code{contar_n_ano_uc()} gera uma tabela contendo o número total de anos em que cada Unidade de Conservação foi amostrada.
}
\examples{
# número de anos em que uma UC foi amostrada - dados brutos
contar_n_ano_uc()

# número de anos em que uma UC foi amostrada - dados filtrados
dados_filtrados <- filtrar_dados(
 dados = monitora_aves_masto_florestal,
 nome_uc == "resex_tapajos_arapiuns"
)

contar_n_ano_uc(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
