% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_obs_sp_ano.R
\name{contar_n_obs_sp_ano}
\alias{contar_n_obs_sp_ano}
\title{Número total de observações para cada espécie e em cada ano}
\usage{
contar_n_obs_sp_ano(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo quatro colunas, ano que contem o ano em que a espécie foi amostrada, \code{nome_sp} contendo o nome da espécie, \code{nome_sp_abv} que contem o nome abreviado da espécie e \code{n} contendo o número de observações realizadas para espécie em cada ano.
}
\description{
A função \code{contar_n_obs_sp_ano()} gera uma tabela contendo o número total de observações para cada espécie e em cada ano.
}
\examples{
# número de observações para cada espécie por ano - dados brutos
contar_n_obs_sp_ano()

# número de observações para cada espécie por ano - dados filtrados
dados_filtrados <- filtrar_dados(
  nome_ucs = "resex_tapajos_arapiuns"
)

contar_n_obs_sp_ano(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
