% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_obs_sp_uc.R
\name{contar_n_obs_sp_uc}
\alias{contar_n_obs_sp_uc}
\title{Número total de observações para cada espécie e em cada Unidade de Conservação}
\usage{
contar_n_obs_sp_uc(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo cinco colunas, \code{nome_uc} que contem o nome da Unidade de Conservação, \code{nome_uc_abv} que contem o nome  abreviado da Unidadade de Conservação, \code{nome_sp} contendo o nome da espécie, \code{nome_sp_abv} que contem o nome abreviado da espécie e \code{n} contendo o número de observações realizadas para espécie em cada ano.
}
\description{
A função \code{contar_n_obs_sp_uc()} gera uma tabela contendo o número total de observações para cada espécie e em cada Unidade de Conservação.
}
\examples{
# número de observações para cada espécie e em cada UC - dados brutos
contar_n_obs_sp_uc()

# número de observações para cada espécie e em cada UC - dados filtrados
dados_filtrados <- filtrar_dados(
  nome_ucs = "resex_tapajos_arapiuns"
)

contar_n_obs_sp_uc(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
