% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_obs_uc.R
\name{contar_n_obs_uc}
\alias{contar_n_obs_uc}
\title{Número total de observações em cada Unidade de Conservação}
\usage{
contar_n_obs_uc(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela funções \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo três colunas, \code{nome_uc} contendo o nome da Unidade de Conservação, \code{nome_uc_abv} que contem o nome abreviado da Unidade de Conservação e \code{n} contendo o número de observações realizadas para cada Unidade de Conservação.
}
\description{
A função \code{contar_n_obs_uc()} gera uma tabela contendo o número total de observações para cada Unidade de Conservação.
}
\examples{
# número de observações para cada UC - dados brutos
contar_n_obs_uc()

# número de observações para cada UC - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_obs_uc(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
