% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{contar_n_obs_validadas}
\alias{contar_n_obs_validadas}
\title{Gera uma tabela com o numero total de observacoes validadas por nivel taxonomico}
\usage{
contar_n_obs_validadas(
  dados = readr::read_rds(file = paste0(here::here(),
    "/inst/extdata/dados_completos.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{contar_n_obs_validadas()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
