% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_uc_ano.R
\name{contar_n_uc_ano}
\alias{contar_n_uc_ano}
\title{Número de Unidades de Conservação amostradas por ano}
\usage{
contar_n_uc_ano(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuraçãoo, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo o número total de Unidades de Conservação amostradas por ano contido na base de dados fornecida.
}
\description{
A função \code{contar_n_uc_ano()} gera o número total de Unidades de Conservação amostradas por ano contidos na base de dados fornecida.
}
\examples{
# número total de UCs em cada ano - dados brutos
contar_n_uc_ano()

# número total de UCs em cada ano - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_uc_ano(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
