% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definir_intervalo_distancia.R
\name{definir_intervalos_distancia}
\alias{definir_intervalos_distancia}
\title{Agrupas as distâncias em intervalos de tamanhos homogêneos ou diferentes}
\usage{
definir_intervalos_distancia(
         dados,
         intervalos_distancia
       )
}
\arguments{
\item{dados}{recebe uma \code{tibble} gerada pela função
\code{\link[=transformar_dados_formato_Distance]{transformar_dados_formato_Distance()}}.}

\item{intervalos_distancia}{recebe um vetor numérico contendo os intervalos
a partir dos quais as distâncias serão agrupadas.}
}
\value{
Retorna a \code{tibble} de entrada contendo duas novas colunas, \code{distbegin}
e \code{distend} que definem o intervalo de agrupamento das distâncias perpendiculares
}
\description{
A função \code{definir_intervalos_distancia()} agrupa as distâncias perpendiculares
nas quais os indivíduos foram observados em intervalos regulares (ex. mesmo
tamanho), ou irregulares (ex. tamanhos diferentes).
}
\examples{
# carregar pacote
library(dplyr)

# gerar dados filtrados
dasy_croc_tap_arap_repeticao <- filtrar_dados(
  nome_uc = "resex_tapajos_arapiuns",
  nome_sps = "dasyprocta_croconota",
  validacao_obs = "especie"
) |>
transformar_dados_formato_Distance()

# definir intervalos de distância (binagem)
dasy_croc_tap_arap_repeticao_binados <- dasy_croc_tap_arap_repeticao |>
  definir_intervalos_distancia(intervalos_distancia = seq(
    from = 0,
    to  = 50,
    by = 1.5
  ))

glimpse(dasy_croc_tap_arap_repeticao_binados)

# definir intervalos de distância diferentes

dasy_croc_tap_arap_repeticao_binados2 <- dasy_croc_tap_arap_repeticao |>
definir_intervalos_distancia(
    intervalos_distancia = c(0, seq(
      from = 1,
      to = 55,
      by = 1.4
    ))
  )

glimpse(dasy_croc_tap_arap_repeticao_binados2)
}
