% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_caracteristicas_area_estudo_taxa_encontro}
\alias{gerar_caracteristicas_area_estudo_taxa_encontro}
\title{Title}
\usage{
gerar_caracteristicas_area_estudo_taxa_encontro(
  dados,
  resultado_selecao_modelos
)
}
\arguments{
\item{resultado_selecao_modelos}{}
}
\description{
Title
}
