% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_dados_completos.R
\name{gerar_dados_completos}
\alias{gerar_dados_completos}
\title{Seleciona, trasforma e renomea as colunas e observações nos dados brutos}
\usage{
gerar_dados_completos(dados)
}
\arguments{
\item{dados}{recebe uma \code{tibble} contendo os dados brutos.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo uma seleção de
colunas transformadas e renomeadas a partir dos dados brutos do Projeto
Monitora Componente Florestal.
}
\description{
A função \code{gerar_dados_completos()} recebe uma \code{tibble} gerada a
partir da função \code{carregar_dados_brutos_xlsx()}, selecionando,
transformando e renomeando suas colunas e observações.
}
\details{
A funcao \code{gerar_dados_completos()} seleciona, transforma e renomeia as
colunas e observações dos dados brutos do Projeto Monitora Componente
Florestal.

A função gera novas colunas contendo o número de vezes que cada estação
amostral foi visitada, a categoria da UC, o nome abreviado das UC's, o nome
abreviado das espécies, o esforço amostral total empregado em cada estação
amostral, o número de observadores e o tempo total de cada censo. Quando
ausentes, as distâncias de trilhas percorridas sao imputadas. Outra
transfomração importante é a inlcusão de linhas para os dias em que houve
amostragem e uma dada espécie não foi observada (ex. distância = \code{NA})
}
\examples{
\dontrun{gerar_dados_completos(dados = dados_brutos)}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
