% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_dados_filtrados}
\alias{gerar_dados_filtrados}
\title{Gerar dados filtrados a partir da UC's e da especie}
\usage{
gerar_dados_filtrados(dados, nome_uc, nome_sp)
}
\arguments{
\item{dados}{recebe o caminho para carregar um arquivo .rds. Por configuracao, carrega os dados filtrados pela Resex Tapajos-Arapiuns e especie \emph{Dsyprocta croconota} do Projeto Monitora Componente Florestal a partir doa arquivo \code{"dados_completos.rds"}.}

\item{nome_uc}{description}

\item{nome_sp}{descricao}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo os dados filtrados do Projeto Monitora Componente Florestal.
}
\description{
A funcao \code{gerar_dados_filtrados()} carrega o arquivo de dados completos, \code{"dados_completos.rds"} e gera um novo arquivo \code{dados_filtrados.rds}.
}
\details{
A funcao \code{gerar_dados_filtrados()} carrega e disponibiliza ao usario os dados filtrados do Projeto Monitora Componente Florestal, contendo ...
}
\examples{
\dontrun{gerar_dados_filtrados()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
