% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_intervalos_confianca_bootstrap.R
\name{gerar_intervalos_confianca_bootstrap}
\alias{gerar_intervalos_confianca_bootstrap}
\title{Estimativa de incerteza para modelos \emph{distance sampling} por \emph{bootstrap}}
\usage{
gerar_intervalos_confianca_bootstrap(
  dados,
  funcao_chave = c("hn", "hr", "unif"),
  termos_ajuste = NULL,
  truncamento = NULL,
  formula = ~1,
  funcao_sumario = c("abundancia", "densidade"),
  nboot = 100,
  nucleos = 1
)
}
\arguments{
\item{dados}{recebe um conjunto de dados filtrados pela função
\code{\link[=filtrar_dados]{filtrar_dados()}} e transformados pela função
\code{\link[=transformar_dados_formato_Distance]{transformar_dados_formato_Distance()}}.}

\item{funcao_chave}{recebe uma dentre três funções utilizadas para modelar a
detectabilidade da espécie: "hn" - Half-normal; "hr" - Hazard-rate; e
"unif" - Uniform.}

\item{termos_ajuste}{recebe um dentre n termos de ajuste: \code{"cos"} - cosseno;
\code{"herm"} - Hermet polynomial; e \code{"poly"} - Polinomial simples. Quando \code{NULL}
diferentes modelos são ajustados para função chave escolhida utilizando
diferentes termos de ajuste. Veja Details.}

\item{truncamento}{recebe uma lista contendo as distâncias de truncagem
de forma numérica (ex. 20), ou como porcentagem (como um caracter, ex. 25\%).
Também pode ser alimentada no formato de lista, com os elementos \code{left} e
\code{right} (ex. list(left = 1, right =20)) se a truncagem a esquerda for
necessária. Por configuração, a distância máxima é utilizada como valor de
truncagem a direita. Quando os dados estão categorizados, a truncagem a
direita é o valor final da última coluna. O valor de truncagem a esquerda
é zero, por configuração.}

\item{formula}{a fórmula para o parâmetro de escala. Para análise CDS manter
a configuração original ~ 1.}

\item{funcao_sumario}{defini se o bootstrap retornará intervalos de confiança
para estimativa de abunância ou de densidade.}

\item{nboot}{número de bootstraps a serem conduzidos.}

\item{nucleos}{número de núcleos do CPU usados para realizar a estimativa do
intervalo de confiança.}
}
\value{
Retornar um vetor contendo o intervalo de confiança e um gáfico de
histograma contendo a distribuição de frequências da estimativa de
abundância ou de densidade e seus respectivos intervalos de confiança.
}
\description{
Conduz um \emph{bootstrap} para modelos de \emph{distance sampling}.
}
\details{
Quando \code{termo_ajuste = NULL} modelos com diferente termos de ajuste serão
ajustados aos dados, a depender da função chave utilizada. Para
\code{funcao_chave = "hn"}, são utilizados os termos de ajuste \code{NULL} (sem termos),
\code{"cos"} e \code{"herm"}; para \code{funcao_chave = "hr"}, são utilizados os termos de
ajuste \code{NULL} (sem termos), \code{"cos"} e \code{"poly"}; e para \code{funcao_chave  = "unif"},
são utilizados so termos de ajuste \code{"cos"} e \code{"poly"}.
}
\examples{
\dontrun{
gerar_intervalos_confianca_bootstrap()
}
}
