% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_tabdin.R
\name{gerar_tabdin}
\alias{gerar_tabdin}
\title{Gera uma tabela dinâmica}
\usage{
gerar_tabdin(dados = monitora_aves_masto_florestal, n_linhas = 1:1000)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burto de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal, \code{monitora_aves_mamiferos_florestal}.}

\item{n_linhas}{fornece o intervalo de linhas a ser gerado. Por configuração, é gerado o intervalo entre as linhas 1 e 1000.}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A função \code{gerar_tabdin()} gera uma tabela dinâmica a partir de uma base de dados de entrada.
}
\details{
A velociadade de geração da tabela dinâmica e o tamanho do intervalo de linhas que pode ser gerado vairam de acordo com a configuração da máquina que executa a função.
}
\examples{
# gerar tabela dinâmica dos dados brutos
gerar_tabdin()

# gerar tabela dinâmica a partir dos dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_sp == c("dasyprocta_croconota", "dasyprocta_iacki")
)
gerar_tabdin(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
