% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_tabdin_dados_brutos}
\alias{gerar_tabdin_dados_brutos}
\title{Gera uma tabela dinamica dos dados brutos}
\usage{
gerar_tabdin_dados_brutos(
  dados = readxl::read_excel(path = paste0(here::here(),
    "/inst/extdata/monitora_masto_aves_2023_04_04.xlsx"), sheet = "dados brutos"),
  n_linhas = 1:1000
)
}
\arguments{
\item{dados}{Descircao}

\item{n_linhas}{Descricao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{gerar_tabdin_dados_brutos()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
