% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_tabdin_dados_completos}
\alias{gerar_tabdin_dados_completos}
\title{Gera uma tabela dinamica dos dados completos}
\usage{
gerar_tabdin_dados_completos(
  dados = readr::read_rds(file = paste0(here::here(), "/data/dados_completos.rds")),
  n_linhas = 1:4500
)
}
\arguments{
\item{dados}{Descircao}

\item{n_linhas}{Descricao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{gerar_tabdin_dados_completos()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
