% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_tabdin_dados_filtrados}
\alias{gerar_tabdin_dados_filtrados}
\title{Gera uma tabela dinamica dos dados selecionados e filtrados por UC e especie}
\usage{
gerar_tabdin_dados_filtrados(
  dados = readr::read_rds(file = paste0(here::here(),
    "/inst/extdata/dados_selecionados.rds")),
  nome_uc = "Resex Tapajos-Arapiuns",
  nome_sp = "Dasyprocta croconota",
  n_linhas = 1:4500
)
}
\arguments{
\item{dados}{Descircao}

\item{nome_uc}{Descricao}

\item{nome_sp}{Descricao}

\item{n_linhas}{Descricao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{gerar_tabdin_dados_filtrados()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
