% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_tabdin_dados_selecionados_distanceR}
\alias{gerar_tabdin_dados_selecionados_distanceR}
\title{Gera uma tabela dinamica dos dados selecionados no formato distance do R sem covariaveis}
\usage{
gerar_tabdin_dados_selecionados_distanceR(
  dados = readr::read_rds(file = paste0(here::here(),
    "/inst/extdata/dados_selecionados_transformados_dist_r.rds")),
  n_linhas = 1:4500
)
}
\arguments{
\item{dados}{Descircao}

\item{n_linhas}{Descricao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{gerar_tabdin_dados_selecionados_distanceR()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
