% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_tabdin_n_ano_uc}
\alias{gerar_tabdin_n_ano_uc}
\title{Gera uma tabela dinamica do numero de anos em que cada UC foi amostrada}
\usage{
gerar_tabdin_n_ano_uc(
  dados = readr::read_rds(file = paste0(here::here(), "/data/n_ano_uc.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{gerar_tabdin_n_ano_uc()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
