% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_tabdin_n_obs_sp_uc_estacao_ano}
\alias{gerar_tabdin_n_obs_sp_uc_estacao_ano}
\title{Gera uma tabela dinamica do numero de observacoes por especie, em cada UC, estacao e ano}
\usage{
gerar_tabdin_n_obs_sp_uc_estacao_ano(
  dados = readr::read_rds(file = paste0(here::here(),
    "/inst/extdata/n_obs_sp_uc_estacao_ano.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{gerar_tabdin_n_obs_sp_uc_estacao_ano()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
