% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_distancia_covariavel.R
\name{plotar_distancia_covariavel}
\alias{plotar_distancia_covariavel}
\title{Plota as distâncias perpendiculares contra uma covariável}
\usage{
plotar_distancia_covariavel(dados, x, y, xlab, ylab)
}
\arguments{
\item{dados}{recebe uma \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}} ou
\code{\link[=transformar_dados_formato_Distance]{transformar_dados_formato_Distance()}}.}

\item{x}{recebe a distância ou a covariável.}

\item{y}{recebe a distância ou a covariável.}

\item{xlab}{recebe o título do eixo x.}

\item{ylab}{recebe o título do eixo y.}
}
\value{
Retorna um gráfico de pontos das distâncias perpendiculares
observadas contra uma covariável.
}
\description{
Breve descrição da função...
}
\examples{
\dontrun{
plotar_distancia_covariavel()
}
}
