% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_distribuicao_distancia_estatico.R
\name{plotar_distribuicao_distancia_estatico}
\alias{plotar_distribuicao_distancia_estatico}
\title{Gráficos interativos das distâncias perpendiculares}
\usage{
plotar_distribuicao_distancia_estatico(
  dados,
  largura_caixa = 1,
  cor = "black",
  preenchimento = "chartreuse4",
  legenda = "auto",
  familia_fonte = "serif",
  tamanho_fonte_titulo_eixos = 12,
  tamanho_fonte_valores_eixos = 11
)
}
\arguments{
\item{dados}{receba uma \code{tibble} gerada a partir da função
\code{\link[=transformar_dados_formato_Distance]{transformar_dados_formato_Distance()}} ou que contenha uma coluna chamada
\code{distance}}

\item{largura_caixa}{recebe um valor que estabelece a largura das colunas
do histograma}
}
\value{
Retorna um objeto do tipo \code{...} contendo um gráfico de pontos
(Cleveland dotplot), de caixa (boxplot) e um histograma, da distribuição
de distâncias perpendiculares.
}
\description{
A funcao \code{plotar_distribuicao_distancia_estatico()} gera três
gráficos, descrevendo a variação das distâncias
perpendiculares observadas.
}
\examples{
\dontrun{plotar_distribuicao_distancia_estatico()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
