% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_distribuicao_distancia_interativo.R
\name{plotar_distribuicao_distancia_interativo}
\alias{plotar_distribuicao_distancia_interativo}
\title{Gráficos interativos das distâncias perpendiculares}
\usage{
plotar_distribuicao_distancia_interativo(dados, largura_caixa = 1)
}
\arguments{
\item{dados}{receba uma \code{tibble} gerada a partir da função
\code{\link[=transformar_dados_formato_Distance]{transformar_dados_formato_Distance()}} ou que contenha uma coluna chamada
\code{distance}}

\item{largura_caixa}{recebe um valor que estabelece a largura das colunas
do histograma}
}
\value{
Retorna um objeto do tipo \code{...} contendo um gráfico de pontos
(Cleveland dotplot), de caixa (boxplot) e um histograma, da distribuição
de distâncias perpendiculares.
}
\description{
A funcao \code{plotar_distribuicao_distancia_interativo()} gera três
gráficos interativos, descrevendo a variação das distâncias
perpendiculares observadas.
}
\examples{
\dontrun{plotar_distribuicao_distancia_interativo()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
