% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{plotar_n_obs_sp_estatico}
\alias{plotar_n_obs_sp_estatico}
\title{Gera graficos de barras estaticos com o numero de observacoes por especie}
\usage{
plotar_n_obs_sp_estatico(
  dados = readr::read_rds(file = paste0(here::here(), "/inst/extdata/n_obs_sp.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{...} contendo graficos de barras estaticos.
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{plotar_n_obs_sp_estatico()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
