% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{plotar_n_obs_validadas_estatico}
\alias{plotar_n_obs_validadas_estatico}
\title{Gera um grafico de barras estatico com o numero de observacoes validadas para cada nivel taxonomico}
\usage{
plotar_n_obs_validadas_estatico(
  dados = readr::read_rds(file = paste0(here::here(),
    "/data/tabela_n_obs_validadas.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{...} contendo graficos de barras interativos.
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{plotar_n_obs_validadas_estatico()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
