% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_resultados_Distance.R
\name{plotar_resultados_Distance}
\alias{plotar_resultados_Distance}
\title{Plota os resultados de abundância ou densidade estimada}
\usage{
plotar_resultados_Distance(
  dados,
  tipo_de_resultado = c("abundância", "densidade"),
  intervalo_confianca = FALSE
)
}
\arguments{
\item{dados}{recebe um objeto do tipo \code{data.frame} gerado pela função
\code{\link[=gerar_resultados_Distance]{gerar_resultados_Distance()}}}

\item{tipo_de_resultado}{...}

\item{intervalo_confianca}{...}
}
\value{
Retorna ...
}
\description{
Descrever o que cada opção retorna
}
\examples{
\dontrun{
plotar_resultados_Distance()
}
}
