% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{selecionar_distancia_truncamento}
\alias{selecionar_distancia_truncamento}
\title{Title}
\usage{
selecionar_distancia_truncamento(
  dados,
  dist_truncamento = list(`25\%` = "25\%", `20\%` = "20\%", `15\%` = "15\%", `10\%` =
    "10\%", `5\%` = "5\%")
)
}
\arguments{
\item{dist_truncamento}{}
}
\description{
Title
}
