% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{selecionar_funcao_deteccao_termo_ajuste}
\alias{selecionar_funcao_deteccao_termo_ajuste}
\title{Title}
\usage{
selecionar_funcao_deteccao_termo_ajuste(dados)
}
\arguments{
\item{dados}{}
}
\description{
Title
}
