% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{testar_bondade_ajuste}
\alias{testar_bondade_ajuste}
\title{Title}
\usage{
testar_bondade_ajuste(
  dados,
  plot = FALSE,
  chisq = FALSE,
  nc = NULL,
  breaks = NULL
)
}
\arguments{
\item{dados}{}
}
\description{
Title
}
