% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{transforma_para_distanceR_com_repeticao_filtra_uc_sp}
\alias{transforma_para_distanceR_com_repeticao_filtra_uc_sp}
\title{Gera uma tabela no formato para analise no pacote Distance do R, sem covariaveis, a partir dos dados selecionados}
\usage{
transforma_para_distanceR_com_repeticao_filtra_uc_sp(
  dados = readr::read_rds(file = paste0(here::here(), "/data/dados_selecionados.rds")),
  nome_uc = "Resex Tapajos-Arapiuns",
  nome_sp = "Dasyprocta croconota"
)
}
\arguments{
\item{dados}{Descricao}
}
\value{
Retorna um objeto do tipo \code{tibble}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{transformar_para_distanceR()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
