% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{transforma_para_dsitanceR_quase_sem_repeticao_filtra_uc_sp}
\alias{transforma_para_dsitanceR_quase_sem_repeticao_filtra_uc_sp}
\title{Title}
\usage{
transforma_para_dsitanceR_quase_sem_repeticao_filtra_uc_sp(
  dados = readr::read_rds(file = paste0(here::here(), "/data/dados_selecionados.rds")),
  nome_uc = "Resex Tapajos-Arapiuns",
  nome_sp = "Dasyprocta croconota"
)
}
\arguments{
\item{nome_sp}{}
}
\description{
Title
}
