% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformar_dados_formato_Distance.R
\name{transformar_dados_formato_Distance}
\alias{transformar_dados_formato_Distance}
\title{Transforma os dados para o formato para análise no pacote \code{Distance}}
\usage{
transformar_dados_formato_Distance(
  dados = monitora_aves_masto_florestal,
  amostras_repetidas = TRUE
)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuraçãoo, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}

\item{amostras_repetidas}{lógico, controla a saída dos dados transformados. Por configuração, retorna os dados com amostragens repetidas em cada estação amostral.

@details
Quando o agrumento \code{amostras_repetidas = FALSE} será selecionado o dia em que foi registrado o maior número de observações em uma estação amostral.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo a base de dados configurada para ajustar os modelos do pacote \code{Distance} a partir da função [ajustar_modelo_distance().
}
\description{
A função \code{transformar_dados_formato_Distance()} transforma a base de dados para o formato de análise do pacote \code{Distance} podendo retornar a base de dados com ou sem amostras repetidas na estações amostrais.
}
\examples{
# carregar pacote
library(dplyr)

# gerar os dados filtrados
dados_filtrados <- filtrar_dados(
  nome_uc = "resex_tapajos_arapiuns",
  nome_sp = "dasyprocta_croconota",
  validacao_obs = "especie"
)

# gerar os dados transformados com repeticao
dados_distance_com_repeticao <- transformar_dados_formato_Distance(dados = dados_filtrados)

glimpse(dados_distance_com_repeticao)

# gerar os dados transformados com repeticao
dados_distance_sem_repeticao <- transformar_dados_formato_Distance(
  dados = dados_filtrados,
  amostras_repetidas = FALSE
)

glimpse(dados_distance_sem_repeticao)
}
