% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{transformar_para_distanceR_covariaveis}
\alias{transformar_para_distanceR_covariaveis}
\title{Gera uma tabela no formato para analise no pacote Distance do R, com duas covariaveis, a partir dos dados selecionados}
\usage{
transformar_para_distanceR_covariaveis(
  dados = readr::read_rds(file = paste0(here::here(),
    "/inst/extdata/dados_selecionados.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{tibble}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{transformar_para_distanceR_covariaveis()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
