% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{carregar_dados_brutos_xlsx}
\alias{carregar_dados_brutos_xlsx}
\title{Carrega os dados brutos originais em formato .xlsx}
\usage{
carregar_dados_brutos_xlsx(caminho, planilha = NULL)
}
\arguments{
\item{dados}{recebe o caminho para o diretório onde se encontra o arquivo .xlsx dos dados brutos.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo os dados brutos do Projeto Monitora Componente Florestal a partir do formato .xlsx.
}
\description{
A funcao \code{carregar_dados_brutos_xlsx()} carrega o arquivo da planilha de dados originais, \code{"monitora_masto_aves_2023_04_04.xlsx"}.
}
\details{
A funcao \code{carregar_dados_brutos_xlsx()} carrega e disponibiliza ao usario os dados brutos do Projeto Monitora Componente Florestal, contendo  as infromacoes sobre as amostragens por distancia (\emph{distance samplig}) de aves e pequenos e medios mamiferos realizadas desde 2014 em 40 Unidades de Conservacao do Brasil.
}
\examples{
\dontrun{carregar_dados_brutos_xlsx("data-raw/dados_brutos.xlsx")}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
