% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{carregar_dados_selecionados}
\alias{carregar_dados_selecionados}
\title{Carrega os dados selecionados a partir das observacoes validadas ao nivel de especie}
\usage{
carregar_dados_selecionados(dados)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{gerar_dados_selecionados(dados = dados_completos)}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
