% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_obs_sp.R
\name{contar_n_obs_sp}
\alias{contar_n_obs_sp}
\title{Número total de observações para cada espécie}
\usage{
contar_n_obs_sp(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burtos de aves e medios e grandes mamiferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo três colunas, \code{nome_sp} contendo o nome da espécie, \code{nome_sp_abv} que contem o nome da espécie abreviado e \code{n} contendo o número de observações realizadas para cada espécie
}
\description{
A função \code{contar_n_obs_sp()} gera uma tabela contendo o número total de observações para cada espécie.
}
\examples{
# número de observações para cada espécie - dados brutos
contar_n_obs_sp()

# número de observações para cada espécie - dados filtrados
dados_filtrados <- filtrar_dados(
  nome_ucs = "resex_tapajos_arapiuns"
)

contar_n_obs_sp(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
