% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{contar_n_obs_sp_uc}
\alias{contar_n_obs_sp_uc}
\title{Gera uma tabela com o numero total de observacoes por especie e por UC}
\usage{
contar_n_obs_sp_uc(
  dados = readr::read_rds(file = paste0(here::here(),
    "/inst/extdata/dados_selecionados.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{}...
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{contar_n_obs_sp_uc()}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
