% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_obs_validadas.R
\name{contar_n_obs_validadas}
\alias{contar_n_obs_validadas}
\title{Número total de observações validadas por nível taxonômico}
\usage{
contar_n_obs_validadas(
         dados = monitora_aves_masto_florestal,
         retorna_tabela = TRUE
       )
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}

\item{retorna_tabela}{lógico, indica se o objeto retornado e uma uma \code{tibble} ou um \code{vetor}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo duas colunas, \code{validacao} contendo os níveis taxonômicos de validação e \code{n} contendo o número de observações validadas para cada nível taxonômico, ou um objeto do tipo \code{vector} nomeado contendo os niveis taxonomicos (nomes) e o número de observações para cada nível taxonômico.
}
\description{
A função \code{contar_n_obs_validadas()} gera uma tabela ou um vetor contendo o número total de observações validadas para cada nível taxonômico.
}
\examples{
# observações para cada nível taxonômico - dados brutos
contar_n_obs_validadas()

# observações para cada nível taxonômico - dados brutos
contar_n_obs_validadas(retorna_tabela = FALSE)

# observações para cada nível taxonômico - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_obs_validadas(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
