% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_sp.R
\name{contar_n_sp}
\alias{contar_n_sp}
\title{Número total de espécies}
\usage{
contar_n_sp(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}
}
\value{
Retorna um objeto do tipo \code{vetor} contendo o número total de espécies contido na base de dados fornecida.
}
\description{
A função \code{contar_n_sp()} gera o número total de espécies contidos na base de dados fornecida.
}
\examples{
# número total de espécies - dados brutos
contar_n_sp()

# número total de espécies para uma UC - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_sp(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
