% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_n_uc.R
\name{contar_n_uc}
\alias{contar_n_uc}
\title{Número de Unidades de Conservação amostradas}
\usage{
contar_n_uc(dados = monitora_aves_masto_florestal)
}
\arguments{
\item{dados}{recebe a \code{tibble} gerada pela função \code{\link[=filtrar_dados]{filtrar_dados()}}. Por configuração, carrega a base de dados brutos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal \code{monitora_aves_masto_florestal}.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo o número total de Unidades de Conservação contido na base de dados fornecida.
}
\description{
A função \code{contar_n_uc()} gera o número total de Unidades de Conservação amostradas contidas na base de dados fornecida.
}
\examples{
# número total de UCs - dados brutos
contar_n_uc()

# número total de UCs - dados filtrados
dados_filtrados <- filtrar_dados(
  nome_sps = "dasyprocta_corconota"
)

contar_n_uc(dados_filtrados)
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
