% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_dados_completos.R
\name{gerar_dados_completos}
\alias{gerar_dados_completos}
\title{Seleciona, trasforma e renomea as colunas e observacoes nos dados brutos}
\usage{
gerar_dados_completos(dados)
}
\arguments{
\item{dados}{recebe uma \code{tibble} contendo os dados brutos.}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo uma selecao de colunas transformadas e renomeadas a partir dos dados brutos do Projeto Monitora Componente Florestal.
}
\description{
A funcao \code{gerar_dados_completos()} recebe uma \code{tibble} gerada a partir da funcao \code{carregar_dados_brutos_xlsx()}, selecionando, transformando e renomeando suas colunas e observacoes.
}
\details{
A funcao \code{gerar_dados_completos()} seleciona, transforma e renomeia as colunas e observacoes dos dados brutos do Projeto Monitora Componente Florestal. A funcao gera novas colunas contendo o numero de vezes que cada estacao amostral foi visitada, a categoria da UC,o nome abreviado das UC's, o nome abreviado das especies, o esforco amostral total empregado em cada estacao amostral, o numero de observadores e o tempo total de cada censo. Quando ausentes, as distancias de trilhas percorridas sao imputadas.
}
\examples{
\dontrun{gerar_dados_completos(dados = dados_brutos)}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
