% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{gerar_dados_completos}
\alias{gerar_dados_completos}
\title{Selecionando, trasforma e renomea as colunas e observacoes nos dados brutos}
\usage{
gerar_dados_completos(dados)
}
\arguments{
\item{dados}{recebe o caminho para carregar um arquivo .rds. Por configuracao, carrega os dados brutos dos Projeto Monitora Componente Florestal a partir do formato .rds}
}
\value{
Retorna um objeto do tipo \code{tibble} contendo uma selecao de colunas transformadas e renomeadas a partir dos dados brutos do Projeto Monitora Componente Florestal.
}
\description{
A funcao \code{gerar_dados_completos()} carrega os dados brutos em formato .rds, selecionando, transformando e renomeando suas colunas.
}
\details{
A funcao \code{gerar_dados_completos()} disponibiliza ao usario os dados brutos do Projeto Monitora Componente Florestal, selecionando, transformando e renomeando suas colunas. (Fornecer mais detalhes sobre as transformacoes operadas).
Avisar que distancias de trilhas percorridas estao sendo imputadas quando os valores estao ausentes.
}
\examples{
\dontrun{gerar_dados_completos(dados = dados_brutos)}
}
\author{
Vitor N. T. Borges-Junior
Luciana A. Fusinatto
}
