% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{plotar_distribuicao_distancia_interativo}
\alias{plotar_distribuicao_distancia_interativo}
\title{Gera tres graficos interativos descrevendo a variacao das distancias perpendiculares observadas}
\usage{
plotar_distribuicao_distancia_interativo(
  dados = readr::read_rds(file = paste0(here::here(),
    "/data/dados_selecionados_transformados_dist_r.rds")),
  largura_caixa = 1
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{...} contendo um grafico de caixa, um grafico de pontos e um histograma da distribuicao de distancias perpendiculares.
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{plotar_distribuicao_distancia_interativo()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
