% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plotar_funcao_deteccao_selecao_distancia_truncamento.R
\name{plotar_funcao_deteccao_selecao_distancia_truncamento}
\alias{plotar_funcao_deteccao_selecao_distancia_truncamento}
\title{Title}
\usage{
plotar_funcao_deteccao_selecao_distancia_truncamento(dados)
}
\arguments{
\item{dados}{breve descrição...}
}
\value{
retorna...
}
\description{
Title
}
\examples{
\dontrun{
plotar_funcao_deteccao_modelos_selecionados()
}
}
