% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_funcoes.R
\name{plotar_n_obs_uc_interativo}
\alias{plotar_n_obs_uc_interativo}
\title{Gera um grafico de barras interativo com o numero de observacoes por UC}
\usage{
plotar_n_obs_uc_interativo(
  dados = readr::read_rds(file = paste0(here::here(), "/inst/extdata/n_obs_uc.rds"))
)
}
\arguments{
\item{dados}{Descircao}
}
\value{
Retorna um objeto do tipo \code{...} contendo graficos de barras interativos.
}
\description{
A funcao \code{} ...
}
\details{
Additional details...
}
\examples{
\dontrun{plotar_n_obs_uc_interativo()}
}
\author{
Vitor N. T. Borges-Junior
Julia L. Luz
Luciana A. Fusinatto
}
