% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecionar_distancia_truncamento.R
\name{selecionar_distancia_truncamento}
\alias{selecionar_distancia_truncamento}
\title{Seleção da melhor distância de truncamento dos dados}
\usage{
selecionar_distancia_truncamento(
  dados,
  dist_truncamento = list(`25\%` = "25\%", `20\%` = "20\%", `15\%` = "15\%", `10\%` =
    "10\%", `5\%` = "5\%")
)
}
\arguments{
\item{dados}{recebe uma \code{tibble} gerada a partir da função
\code{\link[=transformar_dados_formato_Distance]{transformar_dados_formato_Distance()}}}

\item{dist_truncamento}{recebe uma lista contendo as distâncias de truncagem
de forma numérica (ex. 20), ou como porcentagem (como um caracter, ex. 25\%).
Também pode ser alimentada no formato de lista, com os elementos \code{left} e
\code{right} (ex. list(left = 1, right =20)) se a truncagem a esquerda for
necessária. Por configuração, a distância máxima é utilizada como valor de
truncagem a direita. Quando os dados estão categorizados, a truncagem a
direita é o valor final da última coluna. O valor de truncagem a esquerda
é zero, por configuração.}
}
\value{
retorna uma lista contendo...
}
\description{
Breve descrição da função...
}
\examples{
\dontrun{
# gerar dados filtrados
dados_filtrados <- filtrar_dados(
  nome_uc = "resex_tapajos_arapiuns",
  nome_sps = "dasyprocta_croconota",
  validacao_obs = "especie"
)

# gerar os dados transformados com repeticao
dados_distance_com_repeticao <- dados_filtrados |>
  transformar_dados_formato_Distance()


dados_distance_com_repeticao |>
  selecionar_ditancia_truncamento()
}
}
