% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecionar_funcao_deteccao_termo_ajuste.R
\name{selecionar_funcao_deteccao_termo_ajuste}
\alias{selecionar_funcao_deteccao_termo_ajuste}
\title{Seleção de modelos a partir do valor de AIC}
\usage{
selecionar_funcao_deteccao_termo_ajuste(
         ...,
         distancia_categorizada = FALSE,
         delta = FALSE
       )
}
\arguments{
\item{...}{recebe os objetos que aramazenam os modelos a serem sumarizados}

\item{distancia_categorizada}{lógico, se \code{TRUE} informa que as distâncias
foram categorizadas. Por configuração, assume que as distâncias são contínuas}

\item{delta}{adicionar descrição...}
}
\value{
retorna um \code{data.frame} contendo os modelos ordenados de acordo com o
valor de AIC e outras informações adicionais como a presença ou não de
covariáveis, p-valor da estatística utilizada (Chi-quadrado, ou Cramer-von
Mises).
}
\description{
A função \code{selecionar_funcao_deteccao_termo_ajuste()} gera uma tabela de
seleção de modelos para avaliar qual combinação entre função chave e termo
de ajuste melhor se ajustam aos dados.Os modelos são ordenados pelo valor de
AAIC e, portanto não permitem modelos com diferentes distâncias de truncagem
nem diferentes categorização de distâncias.
}
\examples{
\dontrun{
# gerar os dados transformados com repetição
dasy_croc_tap_arap_com_repeticao <- filtrar_dados(
  nome_uc = "resex_tapajos_arapiuns",
  nome_sps = "dasyprocta_croconota",
  validacao_obs = "especie"
) |>
transformar_dados_formato_Distance()

# ajustar modelos com funções chave diferentes
# modelo Half-normal
modelo_hn <- dasy_croc_tap_arap_com_repeticao |>
  ajustar_modelos_Distance(
    funcao_chave = "hn",
    truncamento = 15
  )

# modelo Hazard-rate
modelo_hr <- dasy_croc_tap_arap_com_repeticao |>
  ajustar_modelos_Distance(
    funcao_chave = "hr",
    truncamento = 15
  )

# modelo Uniform
modelo_unif <- dasy_croc_tap_arap_com_repeticao |>
  ajustar_modelos_Distance(
    funcao_chave = "unif",
    truncamento = 15
  )

# gerar a tabela de seleção com o resumo comparativo dos modelos
selecao_funcao_deteccao_termo_ajuste <- selecionar_funcao_deteccao_termo_ajuste(
  modelo_hn$`Sem termo`,
  modelo_hn$Cosseno,
  modelo_hn$`Hermite polinomial`,
  modelo_hr$`Sem termo`,
  modelo_hr$Cosseno,
  modelo_hr$`Polinomial simples`,
  modelo_unif$Cosseno,
  modelo_unif$`Polinomial simples`
)

selecao_funcao_deteccao_termo_ajuste
}
}
